Enable GCM support in eCryptfs
This patch adds support for GCM as a cipher mode in eCryptfs. If a file
is encrypted in this mode, the layout of the lower file will change to
accomodate it. This is due to the need to store auth tags as well as
encrypted data.

The file layout being used is as follows. After the header, the first
extent will hold auth tags. After this auth tag extent, data extents are
written until the first auth tag extent becomes full, at which point
another auth tag extent is created, and subsequent data extents have
auth tags stored in that new extent until it fills up, etc. The last
extent in the file should always be a data extent, except in the case of
a zero length upper file, in which case it is the header. This makes it
so that the lower file will never have a case where it needs to move
extents around or have large gaps in the file.